requirements.txt
scrapy>=2.11
python-dateutil>=2.9
